v1.07.95
Added highlight open orders in order book.
Fixed Mt.Gox My Orders Log tab.
Fixed http timeout and package loss bug.
Added trade type to Bitstamp and BTCChina.
Added auto scroll for Order Book.
Fixed double click on last trades, orders log and order book.
Improved stability and performance.
Added enable and disable all rules function to rules.
Now you able to create cycled rules.
Updated Spanish translation.
Added pro file for Debian, Ubuntu, ect. compilation.

+++ src/historyitem.h
@@ -12,7 +12,7 @@
 
 struct HistoryItem 
 {
-	qint64 date;
+	quint32 date;
 	double volume;
 	double price;
 	QByteArray symbol;
